requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=base.txt base.in
#
amqp==5.0.2               # via kombu
attrs==20.3.0             # via -r base.in, jsonschema
beautifulsoup4==4.9.3     # via -r base.in
cachetools==4.1.1         # via premailer
certifi==2020.12.5        # via requests, sentry-sdk
chardet==3.0.4            # via requests
cssselect==1.1.0          # via premailer
cssutils==1.0.2           # via premailer
django-anymail==8.1       # via -r base.in
django-background-tasks==1.2.5  # via -r base.in
django-ckeditor==6.0.0    # via -r base.in
django-compat==1.0.15     # via django-background-tasks
django-cors-headers==3.5.0  # via -r base.in
django-environ==0.4.5     # via -r base.in
django-extensions==3.1.0  # via -r base.in
django-geojson==3.1.0     # via -r base.in
django-js-asset==1.2.2    # via django-ckeditor
django-jsoneditor==0.1.6  # via -r base.in
django-mail-templated==2.6.5  # via -r base.in
django-nested-admin==3.3.2  # via -r base.in
django-pandas==0.6.2      # via -r base.in
django-phonenumber-field==5.0.0  # via -r base.in
django-recurrence==1.10.3  # via -r base.in
django-redis==4.12.1      # via -r base.in
django-secure==1.0.2      # via -r base.in
django-simple-perms==0.2.8  # via -r base.in
django-split-settings==1.0.1  # via -r base.in
django-tagulous==1.1.0    # via -r base.in
django==2.2.17            # via -r base.in, django-anymail, django-cors-headers, django-geojson, django-phonenumber-field, django-recurrence, django-redis, django-secure, django-simple-perms, django-tagulous, djangorestframework, jsonfield
djangocms-admin-style==2.0.2  # via -r base.in
djangorestframework-jwt==1.11.0  # via -r base.in
djangorestframework==3.12.2  # via -r base.in
gunicorn==20.0.4          # via -r base.in
html2text==2020.1.16      # via -r base.in
html5lib==1.1             # via -r base.in
icalendar==4.0.7          # via -r base.in
idna==2.10                # via requests
importlib-metadata==3.1.1  # via jsonschema, kombu
isodate==0.6.0            # via -r base.in
json-logging-py==0.2      # via -r base.in
jsonfield==3.1.0          # via -r base.in
jsonmerge==1.7.0          # via -r base.in
jsonschema==3.2.0         # via jsonmerge
kombu==5.0.2              # via -r base.in
lxml==4.6.2               # via premailer
numpy==1.19.4             # via pandas
packaging==20.7           # via django-jsoneditor
pandas==1.1.5             # via -r base.in, django-pandas
pendulum==2.1.2           # via -r base.in
phonenumbers==8.12.14     # via -r base.in
pillow==8.0.1             # via -r base.in
premailer==3.7.0          # via -r base.in
psycopg2-binary==2.8.6    # via -r base.in
pyjwt==1.7.1              # via djangorestframework-jwt, twilio
pyparsing==2.4.7          # via packaging
pyrsistent==0.17.3        # via jsonschema
pyshp==2.1.2              # via -r base.in
python-dateutil==2.8.1    # via -r base.in, django-recurrence, icalendar, pandas, pendulum
python-monkey-business==1.0.0  # via django-nested-admin
pytz==2020.4              # via -r base.in, django, django-recurrence, icalendar, pandas, twilio
pytzdata==2020.1          # via pendulum
pyyaml==5.3.1             # via -r base.in
redis==3.5.3              # via -r base.in, django-redis
requests==2.25.0          # via -r base.in, django-anymail, premailer, twilio
scour==0.38.2             # via -r base.in
selenium==3.141.0         # via -r base.in
sentry-sdk==0.19.4        # via -r base.in
serpy==0.3.1              # via -r base.in
simplejson==3.17.2        # via -r base.in
six==1.15.0               # via django-background-tasks, django-compat, django-nested-admin, html5lib, isodate, jsonschema, python-dateutil, python-monkey-business, scour, serpy, transitions, twilio
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.4.1           # via django
transitions==0.8.5        # via -r base.in
twilio==6.50.0            # via -r base.in
unidecode==1.1.1          # via -r base.in
urllib3==1.26.2           # via requests, selenium, sentry-sdk
vine==5.0.0               # via amqp
webencodings==0.5.1       # via html5lib
whitenoise==5.2.0         # via -r base.in
xlsxwriter==1.3.7         # via -r base.in
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
